Move HTTP retryable client to separate file
Copied directly from the metrics client.
